// src/app/interfaces/ventas/mes/ventas-mensuales.ts
import { TipoProductos } from './../tipo-productos';
export interface VentasMensuales {
    TotalVentasAnuales: number;
    Enero: TipoProductos;
    Febrero: TipoProductos;
    Marzo: TipoProductos;
    Abril: TipoProductos;
    Mayo: TipoProductos;
    Junio: TipoProductos;
    Julio: TipoProductos;
    Agosto: TipoProductos;
    Septiembre: TipoProductos;
    Octubre: TipoProductos;
    Noviembre: TipoProductos;
    Diciembre: TipoProductos;
}